requirements.txt
alembic==1.17.1
fastapi==0.121.1
pandas==2.3.3
pydantic==2.12.4
python-dotenv==1.2.1
slowapi==0.1.9
SQLAlchemy==2.0.44
starlette==0.49.0
prometheus_client==0.23.1
python-json-logger=4.0.0